Group by search

+++ backend/src/main/java/com/devsuperior/dsvendas/entities/Sale.java
@@ -20,7 +20,7 @@ public class Sale {
 	private Long id;
 	private Integer visited;
 	private Integer deals;
-	private Integer amount;
+	private Double amount;
 	private LocalDate date;
 
 	@ManyToOne
@@ -30,7 +30,7 @@ public class Sale {
 	public Sale() {
 	}
 
-	public Sale(Long id, Integer visited, Integer deals, Integer amount, LocalDate date, Seller seller) {
+	public Sale(Long id, Integer visited, Integer deals, Double amount, LocalDate date, Seller seller) {
 		this.id = id;
 		this.visited = visited;
 		this.deals = deals;
@@ -63,11 +63,11 @@ public class Sale {
 		this.deals = deals;
 	}
 
-	public Integer getAmount() {
+	public Double getAmount() {
 		return amount;
 	}
 
-	public void setAmount(Integer amount) {
+	public void setAmount(Double amount) {
 		this.amount = amount;
 	}
